requirements.txt
asgiref==3.11.1
Django==4.2.28
django-environ==0.13.0
django-filter==25.1
django-redis==6.0.0
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
numpy==2.4.2
pandas==3.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.2.1
six==1.17.0
sqlparse==0.5.5
